Move fixtures to conftest, use pytest.mark.parametrize on more tests
