+++ tests/test_player_count.py
@@ -3,10 +3,10 @@ import pytest
 from src.parser import QuakeLogParser
 
 
-@pytest.fixture
-def mock_file(tmp_path):
-    file_path = tmp_path / "test.log"
-    file_content = r"""
+@pytest.mark.parametrize(
+    "content",
+    [
+        r"""
 0:00 ------------------------------------------------------------
 0:00 InitGame: \sv_floodProtect\1\sv_maxPing\0\sv_minPing\0\sv_maxRate\10000\sv_minRate\0\sv_hostname\Code Miner Server\g_gametype\0\sv_privateClients\2\sv_maxclients\16\sv_allowDownload\0\dmflags\0\fraglimit\20\timelimit\15\g_maxGameClients\0\capturelimit\8\version\ioq3 1.36 linux-x86_64 Apr 12 2009\protocol\68\mapname\q3dm17\gamename\baseq3\g_needpass\0
 14:00 ClientConnect: 2
@@ -16,10 +16,8 @@ def mock_file(tmp_path):
 20:37 ShutdownGame:
 20:37 ------------------------------------------------------------
 """
-    file_path.write_text(file_content)
-    return file_path
-
-
+    ],
+)
 def test_parse_players_count(mock_file):
     # Arrange
     log_parser = QuakeLogParser(mock_file)
